% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_companyfacts_data.R
\name{describe_sec_companyfacts}
\alias{describe_sec_companyfacts}
\title{Describe Local SEC Company Facts}
\usage{
describe_sec_companyfacts(cik, local_path = NULL)
}
\arguments{
\item{cik}{SEC Central Index Key.}

\item{local_path}{Optional SEC Company Facts storage directory.}
}
\value{
Character scalar narrative.
}
\description{
Describe Local SEC Company Facts
}
