% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_submissions_data.R
\name{describe_sec_submissions}
\alias{describe_sec_submissions}
\title{Describe Local SEC Submissions}
\usage{
describe_sec_submissions(cik, local_path = NULL)
}
\arguments{
\item{cik}{SEC Central Index Key.}

\item{local_path}{Optional SEC submissions storage directory.}
}
\value{
Character scalar narrative.
}
\description{
Describe Local SEC Submissions
}
