% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_data.R
\name{detect_time_gaps_fred}
\alias{detect_time_gaps_fred}
\title{Detect Gaps In Local FRED Data}
\usage{
detect_time_gaps_fred(x, frequency = NULL, local_path = NULL)
}
\arguments{
\item{x}{A FRED \code{data.table}, or a series id.}

\item{frequency}{Optional explicit frequency string.}

\item{local_path}{Optional local storage path if \code{x} is a series id.}
}
\value{
A \code{data.table} of gaps.
}
\description{
Detect Gaps In Local FRED Data
}
