% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okx_candle_data.R
\name{detect_time_gaps_okx_candle}
\alias{detect_time_gaps_okx_candle}
\title{Detect Time Gaps In OKX Candle Data}
\usage{
detect_time_gaps_okx_candle(dt, bar = "4H", tolerance = 1e-04)
}
\arguments{
\item{dt}{A candle \code{data.table}.}

\item{bar}{Candle interval.}

\item{tolerance}{Numeric tolerance for fixed-width gap detection.}
}
\value{
A \code{data.table}.
}
\description{
Detect Time Gaps In OKX Candle Data
}
