% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbstats_data.R
\name{detect_time_gaps_wbstats}
\alias{detect_time_gaps_wbstats}
\title{Detect Gaps In Local World Bank Data}
\usage{
detect_time_gaps_wbstats(
  x = NULL,
  indicator = NULL,
  country = NULL,
  freq = "Y",
  local_path = NULL
)
}
\arguments{
\item{x}{A World Bank \code{data.table}, or \code{NULL} when reading from \code{local_path}.}

\item{indicator}{Optional indicator code when reading local data.}

\item{country}{Optional country code when reading local data.}

\item{freq}{Frequency code.}

\item{local_path}{Optional local storage path.}
}
\value{
A \code{data.table} of gaps.
}
\description{
Detect Gaps In Local World Bank Data
}
