% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okx_candle_data.R
\name{download_candle_from_vm}
\alias{download_candle_from_vm}
\title{Download Candle File from VM}
\usage{
download_candle_from_vm(
  inst_id = NULL,
  bar = "4H",
  crypto_data_path = Sys.getenv("Crypto_Data_Path"),
  exchange = "okx",
  vm_name,
  vm_zone
)
}
\arguments{
\item{inst_id}{Character. Instrument ID (e.g., "ETH-USDT-SWAP").}

\item{bar}{Character. Timeframe tag (e.g., "15m").}

\item{crypto_data_path}{Character. Root data path (default: env Crypto_Data_Path).}

\item{exchange}{Character. Exchange subfolder (default: "okx").}

\item{vm_name}{Character. Remote VM name.}

\item{vm_zone}{Character. Remote VM zone.}
}
\value{
Integer status code from system().
}
\description{
Use gcloud scp to copy a candle .rds file from a VM to local.
}
