% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantage_data.R
\name{get_alphavantage_registry}
\alias{get_alphavantage_registry}
\title{Get Alpha Vantage Series Registry}
\usage{
get_alphavantage_registry(
  registry_path = get_alphavantage_registry_file_path()
)
}
\arguments{
\item{registry_path}{Optional registry JSON path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get Alpha Vantage Series Registry
}
