% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantage_data.R
\name{get_alphavantage_registry_file_path}
\alias{get_alphavantage_registry_file_path}
\title{Get Alpha Vantage Registry File Path}
\usage{
get_alphavantage_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory used for fallback.}
}
\value{
Character scalar path.
}
\description{
Get Alpha Vantage Registry File Path
}
