% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea_data.R
\name{get_bea_registry}
\alias{get_bea_registry}
\title{Get BEA Series Registry}
\usage{
get_bea_registry(registry_path = get_bea_registry_file_path())
}
\arguments{
\item{registry_path}{Optional registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get BEA Series Registry
}
