% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea_data.R
\name{get_bea_registry_file_path}
\alias{get_bea_registry_file_path}
\title{Get BEA Registry File Path}
\usage{
get_bea_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory.}
}
\value{
Character scalar path.
}
\description{
Get BEA Registry File Path
}
