% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_data.R
\name{get_bls_registry}
\alias{get_bls_registry}
\title{Get BLS Registry}
\usage{
get_bls_registry(registry_path = get_bls_registry_file_path())
}
\arguments{
\item{registry_path}{Optional registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get BLS Registry
}
