% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_data.R
\name{get_bls_registry_file_path}
\alias{get_bls_registry_file_path}
\title{Get BLS Registry File Path}
\usage{
get_bls_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory.}
}
\value{
Character scalar path.
}
\description{
Get BLS Registry File Path
}
