% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\name{get_census_registry}
\alias{get_census_registry}
\title{Get Census Series Registry}
\usage{
get_census_registry(registry_path = get_census_registry_file_path())
}
\arguments{
\item{registry_path}{Optional registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get Census Series Registry
}
