% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\name{get_census_registry_file_path}
\alias{get_census_registry_file_path}
\title{Get Census Registry File Path}
\usage{
get_census_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory.}
}
\value{
Character scalar path.
}
\description{
Get Census Registry File Path
}
