% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cftc_cot_data.R
\name{get_cftc_cot_registry}
\alias{get_cftc_cot_registry}
\title{Get CFTC COT Registry}
\usage{
get_cftc_cot_registry(registry_path = get_cftc_cot_registry_file_path())
}
\arguments{
\item{registry_path}{Optional JSON registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get CFTC COT Registry
}
