% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cftc_cot_data.R
\name{get_cftc_cot_registry_file_path}
\alias{get_cftc_cot_registry_file_path}
\title{Get CFTC COT Registry File Path}
\usage{
get_cftc_cot_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory used for fallback.}
}
\value{
Character scalar path.
}
\description{
Get CFTC COT Registry File Path
}
