% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto_derivatives_data.R
\name{get_crypto_derivatives_registry}
\alias{get_crypto_derivatives_registry}
\title{Get Crypto Derivatives Registry}
\usage{
get_crypto_derivatives_registry(
  registry_path = get_crypto_derivatives_registry_file_path()
)
}
\arguments{
\item{registry_path}{Optional JSON registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get Crypto Derivatives Registry
}
