% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_data.R
\name{get_eia_registry}
\alias{get_eia_registry}
\title{Get EIA Series Registry}
\usage{
get_eia_registry(registry_path = get_eia_registry_file_path())
}
\arguments{
\item{registry_path}{Optional JSON registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get EIA Series Registry
}
