% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_data.R
\name{get_eia_registry_file_path}
\alias{get_eia_registry_file_path}
\title{Get EIA Registry File Path}
\usage{
get_eia_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory used for fallback.}
}
\value{
Character scalar path.
}
\description{
Get EIA Registry File Path
}
