% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscaldata_data.R
\name{get_fiscaldata_registry}
\alias{get_fiscaldata_registry}
\title{Get Treasury Fiscal Data Registry}
\usage{
get_fiscaldata_registry(registry_path = get_fiscaldata_registry_file_path())
}
\arguments{
\item{registry_path}{Optional JSON registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get Treasury Fiscal Data Registry
}
