% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscaldata_data.R
\name{get_fiscaldata_registry_file_path}
\alias{get_fiscaldata_registry_file_path}
\title{Get Treasury Fiscal Data Registry File Path}
\usage{
get_fiscaldata_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory used for fallback.}
}
\value{
Character scalar path.
}
\description{
Get Treasury Fiscal Data Registry File Path
}
