% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_data.R
\name{get_fred_registry}
\alias{get_fred_registry}
\title{Get FRED Registry}
\usage{
get_fred_registry(registry_path = get_fred_registry_file_path())
}
\arguments{
\item{registry_path}{Optional JSON registry path.}
}
\value{
\code{data.table}.
}
\description{
Get FRED Registry
}
