% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ishare_data.R
\name{get_ishare_registry}
\alias{get_ishare_registry}
\title{Get iShares Registry}
\usage{
get_ishare_registry(registry_path = get_ishare_registry_file_path())
}
\arguments{
\item{registry_path}{Optional registry JSON path.}
}
\value{
\code{data.table}.
}
\description{
Get iShares Registry
}
