% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ishare_data.R
\name{get_ishare_registry_file_path}
\alias{get_ishare_registry_file_path}
\title{Get iShares Registry File Path}
\usage{
get_ishare_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory used for fallback lookup.}
}
\value{
Character scalar path.
}
\description{
Get iShares Registry File Path
}
