% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphavantage_data.R
\name{get_local_alphavantage_data}
\alias{get_local_alphavantage_data}
\title{Read Local Alpha Vantage Daily Data}
\usage{
get_local_alphavantage_data(symbol, local_path = NULL)
}
\arguments{
\item{symbol}{Market symbol.}

\item{local_path}{Optional Alpha Vantage storage directory.}
}
\value{
A standardized market \code{data.table}, or \code{NULL}.
}
\description{
Read Local Alpha Vantage Daily Data
}
