% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bea_data.R
\name{get_local_bea_data}
\alias{get_local_bea_data}
\title{Read Local BEA Data}
\usage{
get_local_bea_data(series_id, local_path = NULL)
}
\arguments{
\item{series_id}{Registered series identifier.}

\item{local_path}{Optional BEA storage directory.}
}
\value{
A \code{data.table}, or \code{NULL}.
}
\description{
Read Local BEA Data
}
