% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_data.R
\name{get_local_bls_data}
\alias{get_local_bls_data}
\title{Read Local BLS Data}
\usage{
get_local_bls_data(series_id, local_path = NULL)
}
\arguments{
\item{series_id}{BLS series identifier.}

\item{local_path}{Optional BLS storage directory.}
}
\value{
A \code{data.table}, or \code{NULL}.
}
\description{
Read Local BLS Data
}
