% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data.R
\name{get_local_census_data}
\alias{get_local_census_data}
\title{Read Local Census Data}
\usage{
get_local_census_data(series_id, local_path = NULL)
}
\arguments{
\item{series_id}{Registered series identifier.}

\item{local_path}{Optional Census storage directory.}
}
\value{
A \code{data.table}, or \code{NULL}.
}
\description{
Read Local Census Data
}
