% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_local.R
\name{get_local_data_meta}
\alias{get_local_data_meta}
\title{Get Local Data Metadata}
\usage{
get_local_data_meta(local_file_path, fallback_to_file_mtime = TRUE)
}
\arguments{
\item{local_file_path}{Path to a local \code{.rds} data file.}

\item{fallback_to_file_mtime}{Logical. If no sidecar metadata file exists,
use the data file modification time.}
}
\value{
A list of metadata, or \code{NULL}.
}
\description{
Get Local Data Metadata
}
