% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_local.R
\name{get_local_data_utime}
\alias{get_local_data_utime}
\title{Get Local Update Time}
\usage{
get_local_data_utime(local_file_path)
}
\arguments{
\item{local_file_path}{Path to a local \code{.rds} data file.}
}
\value{
POSIXct or \code{NULL}.
}
\description{
Get Local Update Time
}
