% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_data.R
\name{get_local_eia_data}
\alias{get_local_eia_data}
\title{Read Local EIA Data}
\usage{
get_local_eia_data(series_id, local_path = NULL)
}
\arguments{
\item{series_id}{EIA series identifier.}

\item{local_path}{Optional EIA storage directory.}
}
\value{
A \code{data.table}, or \code{NULL}.
}
\description{
Read Local EIA Data
}
