% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiscaldata_data.R
\name{get_local_fiscaldata}
\alias{get_local_fiscaldata}
\title{Read Local Treasury Fiscal Data}
\usage{
get_local_fiscaldata(dataset_id, local_path = NULL)
}
\arguments{
\item{dataset_id}{Registry dataset identifier.}

\item{local_path}{Optional Fiscal Data storage directory.}
}
\value{
A \code{data.table}, or \code{NULL}.
}
\description{
Read Local Treasury Fiscal Data
}
