% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ishare_data.R
\name{get_local_ishare_data}
\alias{get_local_ishare_data}
\title{Get Local iShares Historical Data}
\usage{
get_local_ishare_data(ticker, local_path = NULL)
}
\arguments{
\item{ticker}{ETF ticker.}

\item{local_path}{Optional local storage path.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Local iShares Historical Data
}
