% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ishare_data.R
\name{get_local_ishare_holdings}
\alias{get_local_ishare_holdings}
\title{Get Local iShares Holdings Data}
\usage{
get_local_ishare_holdings(ticker, local_path = NULL)
}
\arguments{
\item{ticker}{ETF ticker.}

\item{local_path}{Optional local storage path.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Local iShares Holdings Data
}
