% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ishare_data.R
\name{get_local_ishare_mega_data}
\alias{get_local_ishare_mega_data}
\title{Get Local iShares Metadata Snapshot}
\usage{
get_local_ishare_mega_data(local_path = NULL)
}
\arguments{
\item{local_path}{Optional local storage path.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Local iShares Metadata Snapshot
}
