% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_data.R
\name{get_local_rss_data}
\alias{get_local_rss_data}
\title{Get Local RSS Data}
\usage{
get_local_rss_data(feed_id, local_path = NULL)
}
\arguments{
\item{feed_id}{Local feed identifier.}

\item{local_path}{Optional local storage path.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Local RSS Data
}
