% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_data.R
\name{get_local_sdmx_data}
\alias{get_local_sdmx_data}
\title{Read Local SDMX Data}
\usage{
get_local_sdmx_data(series_id, local_path = NULL)
}
\arguments{
\item{series_id}{Registry series identifier.}

\item{local_path}{Optional SDMX storage directory.}
}
\value{
A \code{data.table}, or \code{NULL}.
}
\description{
Read Local SDMX Data
}
