% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_companyfacts_data.R
\name{get_local_sec_companyfacts}
\alias{get_local_sec_companyfacts}
\title{Read Local SEC Company Facts}
\usage{
get_local_sec_companyfacts(cik, local_path = NULL)
}
\arguments{
\item{cik}{SEC Central Index Key.}

\item{local_path}{Optional SEC Company Facts storage directory.}
}
\value{
A \code{data.table}, or \code{NULL}.
}
\description{
Read Local SEC Company Facts
}
