% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_submissions_data.R
\name{get_local_sec_submissions}
\alias{get_local_sec_submissions}
\title{Read Local SEC Submissions}
\usage{
get_local_sec_submissions(cik, local_path = NULL)
}
\arguments{
\item{cik}{SEC Central Index Key.}

\item{local_path}{Optional SEC submissions storage directory.}
}
\value{
A \code{data.table}, or \code{NULL}.
}
\description{
Read Local SEC Submissions
}
