% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treasury_data.R
\name{get_local_treasury_rates}
\alias{get_local_treasury_rates}
\title{Get Local Treasury Rates}
\usage{
get_local_treasury_rates(dataset, local_path = NULL)
}
\arguments{
\item{dataset}{Treasury dataset key.}

\item{local_path}{Optional local storage path.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Local Treasury Rates
}
