% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbstats_data.R
\name{get_local_wbstats_data}
\alias{get_local_wbstats_data}
\title{Get Local World Bank Data}
\usage{
get_local_wbstats_data(indicator, country, freq = "Y", local_path = NULL)
}
\arguments{
\item{indicator}{World Bank indicator code.}

\item{country}{Country or aggregate code.}

\item{freq}{Frequency code.}

\item{local_path}{Optional local storage path.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Local World Bank Data
}
