% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_data.R
\name{get_rss_registry}
\alias{get_rss_registry}
\title{Get RSS Registry}
\usage{
get_rss_registry(registry_path = get_rss_registry_file_path())
}
\arguments{
\item{registry_path}{Optional JSON registry path.}
}
\value{
\code{data.table}.
}
\description{
Get RSS Registry
}
