% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_data.R
\name{get_sdmx_registry}
\alias{get_sdmx_registry}
\title{Get SDMX Series Registry}
\usage{
get_sdmx_registry(registry_path = get_sdmx_registry_file_path())
}
\arguments{
\item{registry_path}{Optional JSON registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get SDMX Series Registry
}
