% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx_data.R
\name{get_sdmx_registry_file_path}
\alias{get_sdmx_registry_file_path}
\title{Get SDMX Registry File Path}
\usage{
get_sdmx_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory used for fallback.}
}
\value{
Character scalar path.
}
\description{
Get SDMX Registry File Path
}
