% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_frames_data.R
\name{get_sec_frames_registry}
\alias{get_sec_frames_registry}
\title{Get SEC Frames Registry}
\usage{
get_sec_frames_registry(registry_path = get_sec_frames_registry_file_path())
}
\arguments{
\item{registry_path}{Optional registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get SEC Frames Registry
}
