% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_frames_data.R
\name{get_sec_frames_registry_file_path}
\alias{get_sec_frames_registry_file_path}
\title{Get SEC Frames Registry File Path}
\usage{
get_sec_frames_registry_file_path(config_dir = NULL)
}
\arguments{
\item{config_dir}{Optional configuration directory.}
}
\value{
Character scalar path.
}
\description{
Get SEC Frames Registry File Path
}
