% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_core.R
\name{get_sec_registry}
\alias{get_sec_registry}
\title{Get SEC Company Registry}
\usage{
get_sec_registry(registry_path = get_sec_registry_file_path())
}
\arguments{
\item{registry_path}{Optional JSON registry path.}
}
\value{
A registry \code{data.table}.
}
\description{
Get SEC Company Registry
}
