% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okx_candle_data.R
\name{get_source_data_okx_candle}
\alias{get_source_data_okx_candle}
\title{Get Latest OKX Candle Data}
\usage{
get_source_data_okx_candle(inst_id, bar, limit = 100L, config, tz = "UTC")
}
\arguments{
\item{inst_id}{Instrument identifier.}

\item{bar}{Candle interval.}

\item{limit}{Integer page size.}

\item{config}{OKX API config.}

\item{tz}{Output time zone.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Latest OKX Candle Data
}
