% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eia_data.R
\name{get_source_utime_eia}
\alias{get_source_utime_eia}
\title{Get Latest EIA Series Update Date}
\usage{
get_source_utime_eia(series_id, config = NULL)
}
\arguments{
\item{series_id}{EIA series identifier.}

\item{config}{Optional EIA API configuration.}
}
\value{
A UTC \code{POSIXct}, or \code{NULL}.
}
\description{
Get Latest EIA Series Update Date
}
