% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_companyfacts_data.R
\name{get_source_utime_sec_companyfacts}
\alias{get_source_utime_sec_companyfacts}
\title{Get Latest SEC Company Facts Filing Date}
\usage{
get_source_utime_sec_companyfacts(cik, config = NULL)
}
\arguments{
\item{cik}{SEC Central Index Key.}

\item{config}{Optional SEC configuration.}
}
\value{
A UTC \code{POSIXct}, or \code{NULL}.
}
\description{
Get Latest SEC Company Facts Filing Date
}
