% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_submissions_data.R
\name{get_source_utime_sec_submissions}
\alias{get_source_utime_sec_submissions}
\title{Get Latest SEC Submission Time}
\usage{
get_source_utime_sec_submissions(cik, config = NULL)
}
\arguments{
\item{cik}{SEC Central Index Key.}

\item{config}{Optional SEC configuration.}
}
\value{
A UTC-compatible \code{POSIXct}, or \code{NULL}.
}
\description{
Get Latest SEC Submission Time
}
