% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbstats_data.R
\name{get_source_utime_wbstats}
\alias{get_source_utime_wbstats}
\title{Get Fallback Source Update Time for wbstats Data}
\usage{
get_source_utime_wbstats(freq = "Y", tz = "UTC")
}
\arguments{
\item{freq}{Frequency code: \code{"Y"}, \code{"Q"}, or \code{"M"}.}

\item{tz}{Time zone used for the inferred timestamp.}
}
\value{
POSIXct.
}
\description{
Get Fallback Source Update Time for wbstats Data
}
