% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbstats_data.R
\name{get_wbstats_registry}
\alias{get_wbstats_registry}
\title{Get World Bank Registry}
\usage{
get_wbstats_registry(registry_path = get_wbstats_registry_file_path())
}
\arguments{
\item{registry_path}{Optional JSON registry path.}
}
\value{
\code{data.table}.
}
\description{
Get World Bank Registry
}
