% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantmod_data.R
\name{get_yahoofinance_registry}
\alias{get_yahoofinance_registry}
\title{Get Yahoo Finance Registry}
\usage{
get_yahoofinance_registry(
  registry_path = get_yahoofinance_registry_file_path()
)
}
\arguments{
\item{registry_path}{Optional registry JSON path.}
}
\value{
\code{data.table}.
}
\description{
Get Yahoo Finance Registry
}
