% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_specs.R
\name{list_source_specs}
\alias{list_source_specs}
\title{List Source Specifications}
\usage{
list_source_specs()
}
\value{
Named list of \code{investdatar_source_spec} objects.
}
\description{
List Source Specifications
}
