% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_data.R
\name{sync_all_fred_registry_data}
\alias{sync_all_fred_registry_data}
\title{Synchronize All FRED Series In The Registry}
\usage{
sync_all_fred_registry_data(
  registry = get_fred_registry(),
  config = NULL,
  local_path = NULL,
  from_server = FALSE,
  tz = "America/Chicago"
)
}
\arguments{
\item{registry}{Optional FRED registry table.}

\item{config}{Optional FRED API config.}

\item{local_path}{Optional local storage path.}

\item{from_server}{Logical. If \code{TRUE}, use FRED server-reported update time.}

\item{tz}{Time zone used for source update time inference.}
}
\value{
Summary \code{data.table}.
}
\description{
Synchronize All FRED Series In The Registry
}
