% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbstats_data.R
\name{sync_all_wbstats_registry_data}
\alias{sync_all_wbstats_registry_data}
\title{Synchronize All World Bank Registry Series}
\usage{
sync_all_wbstats_registry_data(
  registry = get_wbstats_registry(),
  local_path = NULL,
  ...
)
}
\arguments{
\item{registry}{Optional World Bank registry table.}

\item{local_path}{Optional local storage path.}

\item{...}{Passed to \code{sync_local_wbstats_data()}.}
}
\value{
Summary \code{data.table}.
}
\description{
Synchronize All World Bank Registry Series
}
