% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binance_data.R
\name{get_local_binance_klines}
\alias{get_local_binance_klines}
\title{Get Local Binance Kline Data}
\usage{
get_local_binance_klines(
  symbol = "ETHUSDT",
  interval = "1m",
  local_path = NULL
)
}
\arguments{
\item{symbol}{Trading pair symbol.}

\item{interval}{Candlestick interval.}

\item{local_path}{Optional Binance storage path.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Local Binance Kline Data
}
