% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_data.R
\name{get_local_FRED_data}
\alias{get_local_FRED_data}
\title{Get Local FRED Data}
\usage{
get_local_FRED_data(series_id, local_path = NULL)
}
\arguments{
\item{series_id}{FRED series identifier.}

\item{local_path}{Optional local storage path.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Local FRED Data
}
