% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okx_candle_data.R
\name{get_local_okx_candle}
\alias{get_local_okx_candle}
\title{Get Local OKX Candle Data}
\usage{
get_local_okx_candle(inst_id, bar, local_path = NULL)
}
\arguments{
\item{inst_id}{Instrument identifier.}

\item{bar}{Candle interval.}

\item{local_path}{Optional OKX storage path.}
}
\value{
\code{data.table} or \code{NULL}.
}
\description{
Get Local OKX Candle Data
}
