% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred_data.R
\name{get_source_utime_fred}
\alias{get_source_utime_fred}
\title{Get FRED Series Last Update Time}
\usage{
get_source_utime_fred(series_id, config)
}
\arguments{
\item{series_id}{Character. FRED series ID.}

\item{config}{List with fields: api_key, url, mode.}
}
\value{
POSIXct (UTC).
}
\description{
Return the last update time of a FRED series as POSIXct in UTC.
}
