% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okx_candle_data.R
\name{get_source_utime_okx_candle}
\alias{get_source_utime_okx_candle}
\title{Get Last Completed Candle Time}
\usage{
get_source_utime_okx_candle(bar, tz)
}
\arguments{
\item{bar}{Character. OKX timeframe tag (e.g., "1m", "15m", "4h", "1d").}

\item{tz}{Character. IANA timezone (e.g., "UTC").}
}
\value{
POSIXct in the given timezone.
}
\description{
Floor current time to the last completed candle start for the given bar.
}
