% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_data.R
\name{get_source_utime_rss}
\alias{get_source_utime_rss}
\title{Get RSS Feed Last Update Time}
\usage{
get_source_utime_rss(feed_id, url, parser = c("plain", "gdpnow"))
}
\arguments{
\item{feed_id}{Local feed identifier.}

\item{url}{RSS feed URL.}

\item{parser}{Parsing strategy. Currently supports \code{"plain"} and \code{"gdpnow"}.}
}
\value{
POSIXct or \code{NULL}.
}
\description{
Get RSS Feed Last Update Time
}
