% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rss_data.R
\name{sync_all_rss_registry_data}
\alias{sync_all_rss_registry_data}
\title{Synchronize All RSS Registry Feeds}
\usage{
sync_all_rss_registry_data(registry = get_rss_registry(), local_path = NULL)
}
\arguments{
\item{registry}{Optional RSS registry table.}

\item{local_path}{Optional local storage path.}
}
\value{
Summary \code{data.table}.
}
\description{
Synchronize All RSS Registry Feeds
}
