% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantmod_data.R
\name{sync_all_yahoofinance_registry_data}
\alias{sync_all_yahoofinance_registry_data}
\title{Synchronize All Yahoo Finance Tickers In The Registry}
\usage{
sync_all_yahoofinance_registry_data(
  from,
  to,
  registry = get_yahoofinance_registry(),
  local_path = NULL,
  src = "yahoo"
)
}
\arguments{
\item{from}{Start date passed to \code{quantmod::getSymbols()}.}

\item{to}{End date passed to \code{quantmod::getSymbols()}.}

\item{registry}{Optional Yahoo Finance registry table.}

\item{local_path}{Optional local storage path.}

\item{src}{quantmod source, default \code{"yahoo"}.}
}
\value{
Summary \code{data.table}.
}
\description{
Synchronize All Yahoo Finance Tickers In The Registry
}
